+++ pcs/lib/commands/stonith.py
@@ -1,3 +1,7 @@
+from typing import (
+    Optional,
+)
+
 from pcs.lib import reports
 from pcs.lib.cib import resource
 from pcs.lib.cib.resource.common import are_meta_disabled
@@ -6,13 +10,14 @@ from pcs.lib.commands.resource import (
     _ensure_disabled_after_wait,
     resource_environment
 )
+from pcs.lib.env import LibraryEnvironment
 from pcs.lib.errors import LibraryError
 from pcs.lib.pacemaker.live import (
     FenceHistoryCommandErrorException,
     fence_history_cleanup,
     fence_history_text,
     fence_history_update,
-    is_fence_history_supported,
+    is_fence_history_supported_management,
 )
 from pcs.lib.pacemaker.values import validate_id
 from pcs.lib.resource_agent import find_valid_stonith_agent_by_name as get_agent
@@ -162,51 +167,54 @@ def create_in_group(
             put_after_adjacent,
         )
 
-def history_get_text(env, node=None):
+def history_get_text(env: LibraryEnvironment, node: Optional[str] = None):
     """
     Get full fencing history in plain text
 
-    LibraryEnvironment env
-    string node -- get history for the specified node or all nodes if None
+    env
+    node -- get history for the specified node or all nodes if None
     """
-    if not is_fence_history_supported():
+    runner = env.cmd_runner()
+    if not is_fence_history_supported_management(runner):
         raise LibraryError(reports.fence_history_not_supported())
 
     try:
-        return fence_history_text(env.cmd_runner(), node)
+        return fence_history_text(runner, node)
     except FenceHistoryCommandErrorException as e:
         raise LibraryError(
             reports.fence_history_command_error(str(e), "show")
         )
 
-def history_cleanup(env, node=None):
+def history_cleanup(env: LibraryEnvironment, node: Optional[str] = None):
     """
     Clear fencing history
 
-    LibraryEnvironment env
-    string node -- clear history for the specified node or all nodes if None
+    env
+    node -- clear history for the specified node or all nodes if None
     """
-    if not is_fence_history_supported():
+    runner = env.cmd_runner()
+    if not is_fence_history_supported_management(runner):
         raise LibraryError(reports.fence_history_not_supported())
 
     try:
-        return fence_history_cleanup(env.cmd_runner(), node)
+        return fence_history_cleanup(runner, node)
     except FenceHistoryCommandErrorException as e:
         raise LibraryError(
             reports.fence_history_command_error(str(e), "cleanup")
         )
 
-def history_update(env):
+def history_update(env: LibraryEnvironment):
     """
     Update fencing history in a cluster (sync with other nodes)
 
-    LibraryEnvironment env
+    env
     """
-    if not is_fence_history_supported():
+    runner = env.cmd_runner()
+    if not is_fence_history_supported_management(runner):
         raise LibraryError(reports.fence_history_not_supported())
 
     try:
-        return fence_history_update(env.cmd_runner())
+        return fence_history_update(runner)
     except FenceHistoryCommandErrorException as e:
         raise LibraryError(
             reports.fence_history_command_error(str(e), "update")